Move types to a separate file
Related to #1

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/bmordue/miniap/issues/1?shareId=XXXX-XXXX-XXXX-XXXX).
